Create README.md
We have received data from clients of a bank many of whom have left. Our goal is to find those factors that are affecting whether someone will leave or not.
finally we have to build a model that predicts with the greatest possible accuracy if someone will exit or not.